requirements.txt
# 核心依赖
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
python-dotenv>=1.0.0

# 数据库
sqlalchemy>=2.0.25
alembic>=1.13.1
pymysql>=1.1.0
aiomysql>=0.2.0

# 向量数据库
chromadb>=0.4.22
sentence-transformers>=2.2.2

# LLM相关
openai>=1.10.0
langchain>=0.1.4
langchain-community>=0.0.16
langchain-openai>=0.0.5

# 文档处理
pypdf>=4.0.1
python-docx>=1.1.0

# 工具库
numpy>=1.26.3
pandas>=2.2.0
requests>=2.31.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# 日志
loguru>=0.7.2

# Web界面
gradio>=4.0.0
streamlit>=1.30.0
